python-server/requirements.txt
flask
matplotlib
pandas
psycopg2-binary
sqlalchemy
uuid